Binance_bot_trade/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bitarray==3.1.0
bitcoinlib==0.7.2
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
colorama==0.4.6
cytoolz==1.0.1
dateparser==1.2.1
dnspython==2.7.0
ecdsa==0.19.0
eth_abi==5.2.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
hexbytes==1.3.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lunardate==0.2.2
MarkupSafe==3.0.2
mnemonic==0.20
multidict==6.1.0
numpy==2.2.3
pandas==2.2.3
parsimonious==0.10.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
PyNaCl==1.5.0
pyparsing==3.2.1
python-binance==1.0.27
python-dateutil==2.9.0.post0
pytz==2025.1
pyunormalize==16.0.0
pywin32==308
regex==2024.11.6
requests==2.32.3
requests-sse==0.3.2
rlp==4.1.0
rsa==4.9
schedule==1.2.2
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.38
stellar-sdk==12.1.0
toml==0.10.2
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
web3==7.8.0
websockets==13.1
Werkzeug==3.1.3
xdrlib3==0.1.1
yarl==1.18.3